emission_factors_api/requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.2
pydantic-settings==2.1.0
numpy==1.24.3
scipy==1.10.1
scikit-learn==1.2.2
sentence-transformers==2.2.2
python-dotenv==1.0.0
python-multipart==0.0.6
faiss-cpu==1.7.4
boto3==1.34.11
botocore==1.34.11
requests==2.31.0 